+++ packages/slate/src/core/index.ts
@@ -2,4 +2,5 @@ export * from './apply'
 export * from './get-dirty-paths'
 export * from './get-fragment'
 export * from './normalize-node'
+export * from './on-error'
 export * from './should-normalize'
