+++ packages/slate/src/interfaces/index.ts
@@ -1,5 +1,6 @@
 export * from './editor'
 export * from './element'
+export * from './error-logger'
 export * from './location'
 export * from './node'
 export * from './operation'
